Add initial unit tests
