+++ _socketsGenerator/generateCiCredentials.js
