+++ _socketsGenerator/machines.json
