+++ _socketsGenerator/package.json
@@ -4,7 +4,13 @@
   "description": "Syncano sockets repository",
   "main": "script.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "sockets": "node createSockets.js",
+    "account": "node createAccount.js",
+    "e2e": "node createAccount.js && node test.js",
+    "start": "node createAccount.js && node createSockets.js && node test.js",
+    "test": "mocha --require babel-register",
+    "check-coverage": "nyc npm run test",
+    "test:watch": "npm t -- -w"
   },
   "repository": {
     "type": "git",
@@ -23,22 +29,24 @@
   },
   "homepage": "https://github.com/Syncano/sockets#readme",
   "devDependencies": {
-    "form-data": "^2.1.2",
-    "rimraf": "^2.5.4",
-     "archiver": "^1.2.0",
+    "archiver": "^1.2.0",
     "axios": "^0.15.3",
+    "babel-cli": "6.18.0",
+    "babel-preset-latest": "6.16.0",
+    "babel-register": "^6.18.0",
+    "chai": "^3.5.0",
+    "chai-as-promised": "^6.0.0",
     "dedent": "^0.6.0",
+    "form-data": "^2.1.2",
     "fs": "0.0.1-security",
     "json2yaml": "^1.1.0",
     "lodash": "^4.17.2",
     "mkdirp": "^0.5.1",
+    "mocha": "^3.2.0",
+    "nyc": "^10.0.0",
+    "rimraf": "^2.5.4",
+    "sinon": "^1.17.6",
     "syncano": "^1.0.28",
     "write-yaml": "^0.2.2"
-  },
-   "scripts": {
-    "sockets": "node createSockets.js",
-    "account": "node createAccount.js",
-    "test": "node createAccount.js && node test.js",
-    "start": "node createAccount.js && node createSockets.js && node test.js"
   }
 }
