+++ _socketsGenerator/script.js
